move the nkfdNormalizations method to the AbstractNameStandardizer class
